Fix bug where +- grades are grouped together

+++ src/main/java/seedu/modtrek/model/module/Grade.java
@@ -100,6 +100,6 @@ public class Grade implements Comparable<Grade> {
      */
     @Override
     public int compareTo(Grade o) {
-        return (int) (o.toPoints() - this.toPoints());
+        return (int) o.toPoints() - (int) this.toPoints();
     }
 }
